requirements.txt
torch>=1.12.0
transformers>=4.25.0
Pillow>=9.3.0
numpy>=1.23.5
pandas>=1.5.2
tqdm>=4.64.1
flask>=2.2.2
python-dotenv>=0.21.0
scikit-learn>=1.2.0
nltk>=3.8.0
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.20.8
requests>=2.28.1
matplotlib>=3.6.2
seaborn>=0.12.1
opencv-python>=4.6.0
sentence-transformers>=2.2.2 